requirements.txt
Flask==3.0.0
requests==2.31.0
moviepy==1.0.3
Pillow==10.1.0
gTTS==2.4.0
python-dotenv==1.0.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
numpy==1.24.3
edge-tts==6.1.9

# ElevenLabs API للصوت الاحترافي
elevenlabs==1.0.0

# معالجة الصوت المحلية
pydub==0.25.1
scipy==1.11.4
noisereduce==3.0.0
pyloudnorm==0.1.1

# تحسين الفيديو المحلي
opencv-python==4.8.1.78
scikit-image==0.22.0

# توليد الصور والفيديو بالذكاء الاصطناعي
openai==1.12.0
stability-sdk==0.8.5
replicate==0.22.0

# أدوات إضافية
aiofiles==23.2.1
psutil==5.9.6
cachetools==5.3.2
concurrent-futures>=3.1.1; python_version < '3.2'
ffmpeg-python==0.2.0
colorama==0.4.6
tqdm==4.66.1
